chore: improving test structure
