+++ composer.json
@@ -1,5 +1,5 @@
 {
-    "name": "aw-studio/laravel-deeplable",
+    "name": "larshanskrause/laravel-deeplable",
     "license": "MIT",
     "authors": [
         {
