+++ phpunit.xml
@@ -12,7 +12,7 @@
          verbose="true"
 >
     <testsuites>
-        <testsuite name="Fjord Test Suite">
+        <testsuite name="Tests">
             <directory suffix="Test.php">./tests</directory>
         </testsuite>
     </testsuites>
