+++ tests/AstrotomicTranslatorTest.php
@@ -2,36 +2,19 @@
 
 namespace Tests;
 
-use Astrotomic\Translatable\Contracts\Translatable as TranslatableContract;
-use Astrotomic\Translatable\Translatable;
 use Astrotomic\Translatable\TranslatableServiceProvider;
 use AwStudio\Deeplable\Deepl;
 use AwStudio\Deeplable\Translators\AstrotomicTranslator;
-use Illuminate\Database\Eloquent\Model;
 use Mockery;
-use Orchestra\Testbench\TestCase;
+use Tests\TestSupport\DummyPost;
 
 class AstrotomicTranslatorTest extends TestCase
 {
-    public function setUp(): void
-    {
-        parent::setUp();
-    }
-
     protected function getPackageProviders($app)
     {
         return [
             TranslatableServiceProvider::class,
         ];
-    }
-
-    protected function getEnvironmentSetUp($app)
-    {
-        $config = $app['config'];
-        $config->set('translatable.fallback_locale', 'en');
-        $config->set('translatable.locales', [
-            'en', 'de',
-        ]);
     }
 
     public function testItTranslatesModelAttributes()
@@ -40,7 +23,7 @@ class AstrotomicTranslatorTest extends TestCase
         $api->shouldReceive('translate')->andReturn('Hallo Welt');
         $translator = new AstrotomicTranslator($api);
 
-        $post = new DummyTranslatablePost([
+        $post = new DummyPost([
             'en' => ['title' => 'Hello World'],
             'de' => ['title' => 'Foo'],
         ]);
@@ -52,7 +35,7 @@ class AstrotomicTranslatorTest extends TestCase
 
         $this->app->setLocale('en');
 
-        $post = new DummyTranslatablePost([
+        $post = new DummyPost([
             'en' => ['title' => 'Hello World'],
             'de' => ['title' => 'Foo'],
         ]);
@@ -69,7 +52,7 @@ class AstrotomicTranslatorTest extends TestCase
         $api->shouldReceive('translate')->andReturn('Hallo Welt');
         $translator = new AstrotomicTranslator($api);
 
-        $post = new DummyTranslatablePost([
+        $post = new DummyPost([
             'en' => ['title' => 'Hello World'],
         ]);
 
@@ -82,7 +65,7 @@ class AstrotomicTranslatorTest extends TestCase
         $api->shouldReceive('translate')->andReturn('Hallo Welt');
         $translator = new AstrotomicTranslator($api);
 
-        $post = new DummyTranslatablePost([
+        $post = new DummyPost([
             'en' => ['title' => 'Hello World'],
             'de' => ['title' => 'Foo'],
         ]);
@@ -92,7 +75,7 @@ class AstrotomicTranslatorTest extends TestCase
         $this->assertSame($post->title, 'Foo');
 
         $this->app->setLocale('en');
-        $post = new DummyTranslatablePost([
+        $post = new DummyPost([
             'en' => ['title' => 'Hello World'],
             'de' => ['title' => 'Foo'],
         ]);
@@ -100,21 +83,4 @@ class AstrotomicTranslatorTest extends TestCase
         $this->app->setLocale('de');
         $this->assertSame($post->title, 'Foo');
     }
-}
-
-class DummyTranslatablePostTranslation extends Model
-{
-    public $table = 'post_translations';
-    protected $fillable = ['title'];
-    public $timestamps = false;
-}
-
-class DummyTranslatablePost extends Model implements TranslatableContract
-{
-    use Translatable;
-
-    public $table = 'posts';
-    protected $translationModel = DummyTranslatablePostTranslation::class;
-    protected $translatedAttributes = ['title'];
-    public $timestamps = false;
 }
